+++ R/plot_ffs.R
@@ -4,13 +4,17 @@
 #' the standard errors from cross validation.
 #' Marked points show the best model from each number of variables until a further variable
 #' could not improve the results.
+#' If type=="selected", the contribution of the selected variables to the model
+#' performance is shown.
 #' @param ffs_model Result of a forward feature selection see \code{\link{ffs}}
+#' @param plotType character. Either "all" or "selected"
 #' @param palette A color palette
 #' @param reverse Character. Should the palette be reversed?
 #' @param marker Character. Color to mark the best models
 #' @param size Numeric. Size of the points
 #' @param lwd Numeric. Width of the error bars
 #' @param pch Numeric. Type of point marking the best models
+#' @param ... Further arguments for base plot if type="selected"
 #' @author Marvin Ludwig and Hanna Meyer
 #' @seealso \code{\link{ffs}}, \code{\link{bss}}
 #' @examples
@@ -18,21 +22,46 @@
 #' data(iris)
 #' ffsmodel <- ffs(iris[,1:4],iris$Species)
 #' plot_ffs(ffsmodel)
+#' #plot performance of selected variables only:
+#' plot_ffs(ffsmodel,plotType="selected)
 #'}
 #' @export plot_ffs
 #' @aliases plot_ffs plot_bss
 
 
-plot_ffs <- function(ffs_model,palette=rainbow,reverse=FALSE, marker="black",size=1.5,lwd=0.5,
-                     pch=21){
+plot_ffs <- function(ffs_model,plotType="all",palette=rainbow,reverse=FALSE, marker="black",size=1.5,lwd=0.5,
+                     pch=21,...){
   metric <- ffs_model$metric
+  if (is.null(ffs_model$type)){
+    ffs_model$type <- "ffs"
+  }
+  if(ffs_model$type=="bss"&plotType=="selected"){
+    type <- "all"
+    print("warning: type must be 'all' for a bss model")
+  }
+  if (plotType=="selected"){
+    labels <- paste0(ffs_model$selectedvars[1],
+                     "+",ffs_model$selectedvars[2])
+    for (i in 3:length(ffs_model$selectedvars)){
+      labels <- c(labels,paste0("+",ffs_model$selectedvars[i]))
+    }
+    plot(ffs_model$selectedvars_perf,xaxt="n",xlab="",
+         ylab=metric,
+         ylim=c(min(ffs_model$selectedvars_perf-ffs_model$selectedvars_perf_SE,na.rm=TRUE),
+                max(ffs_model$selectedvars_perf+ffs_model$selectedvars_perf_SE,na.rm=TRUE)),
+           ...)
+    segments(1:(length(ffs_model$selectedvars)-1),
+             ffs_model$selectedvars_perf-ffs_model$selectedvars_perf_SE,
+             1:(length(ffs_model$selectedvars)-1),
+             ffs_model$selectedvars_perf+ffs_model$selectedvars_perf_SE)
+    axis(1,at=1:(length(ffs_model$selectedvars)-1),labels=labels,las=2)
+  }else{
+
+
   output_df <- ffs_model$perf_all
   output_df$run <- seq(nrow(output_df))
   names(output_df)[which(names(output_df)==metric)] <- "value"
 
-  if (is.null(ffs_model$type)){
-    ffs_model$type <- "ffs"
-  }
   if (ffs_model$type=="bss"){
     bestmodels <- output_df$run[which(output_df$value==ffs_model$selectedvars_perf)]
   }else{
@@ -87,5 +116,7 @@ plot_ffs <- function(ffs_model,palette=rainbow,reverse=FALSE, marker="black",siz
       ggplot2::scale_colour_manual(values = cols, name = "variables")
 
   }
+
   return(p)
+}
 }